# donation-app/src/main/resources/openapi/api.yaml
openapi: 3.0.3
info:
  title: DonationSpace API
  description: Спецификация API для реактивного приложения донатов DonationSpace уровня Праймал.
  version: 1.1.0
paths:
  /api/auth/register:
    post:
      tags:
        - Authentication
      summary: Регистрация нового пользователя
      operationId: register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRequest'
      responses:
        '201':
          description: Успешная регистрация
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
        '400':
          description: Ошибка валидации
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '409':
          description: Пользователь уже существует
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/login:
    post:
      tags:
        - Authentication
      summary: Авторизация пользователя (Вход)
      operationId: login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AuthRequest'
      responses:
        '200':
          description: Успешная авторизация
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Неверные учетные данные
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/auth/login/verify-mfa:
    post:
      tags:
        - Authentication
      summary: Второй шаг авторизации (Верификация MFA)
      operationId: verifyMfaAndLogin
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MfaVerificationRequest'
      responses:
        '200':
          description: Успешная верификация и вход
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401':
          description: Неверный код безопасности
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    AuthRequest:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: user@example.com
        password:
          type: string
          minLength: 6
          example: strongPassword123
    UserResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        uuid:
          type: string
          format: uuid
        email:
          type: string
        role:
          type: string
    LoginResponse:
      type: object
      properties:
        mfaRequired:
          type: boolean
        mfaType:
          type: string
        token:
          type: string
        email:
          type: string
        uuid:
          type: string
          format: uuid
    MfaVerificationRequest:
      type: object
      required:
        - email
        - code
      properties:
        email:
          type: string
        code:
          type: string
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string